convert Integer to Long in DTO

+++ src/main/java/com/farshad/tracker/simpleissuetracker/base/dto/DeveloperPointsDto.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class DeveloperPointsDto {
-    private Integer developerId;
+    private Long developerId;
     private Integer points;
 }
